+++ Navigation/AbstractNavigationNode.php
@@ -11,7 +11,7 @@
 
 namespace WBW\Bundle\CoreBundle\Navigation;
 
-use WBW\Library\Core\Model\Node\AbstractNode;
+use WBW\Library\Core\Sorting\AlphabeticalTreeNodeInterface;
 
 /**
  * Abstract navigation node.
@@ -20,7 +20,7 @@ use WBW\Library\Core\Model\Node\AbstractNode;
  * @package WBW\Bundle\CoreBundle\Navigation
  * @abstract
  */
-abstract class AbstractNavigationNode extends AbstractNode implements NavigationInterface {
+abstract class AbstractNavigationNode implements NavigationInterface, AlphabeticalTreeNodeInterface {
 
     /**
      * Active ?
@@ -44,11 +44,39 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
     private $icon;
 
     /**
+     * Id.
+     *
+     * @var string
+     */
+    private $id;
+
+    /**
+     * Index.
+     *
+     * @var array
+     */
+    private $index;
+
+    /**
      * Matcher.
      *
      * @var string
      */
     private $matcher;
+
+    /**
+     * Navigation nodes.
+     *
+     * @var AbstractNavigationNode[]
+     */
+    private $nodes;
+
+    /**
+     * Parent.
+     *
+     * @var AbstractNavigationNode
+     */
+    private $parent;
 
     /**
      * Target.
@@ -80,14 +108,41 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
      * @param string $matcher The matcher.
      */
     protected function __construct($name, $icon = null, $uri = null, $matcher = self::NAVIGATION_MATCHER_URL) {
-        parent::__construct($name);
         $this->setActive(false);
         $this->setEnable(false);
         $this->setIcon($icon);
+        $this->setId($name);
+        $this->setIndex([]);
         $this->setMatcher($matcher);
+        $this->setNodes([]);
+        $this->setParent(null);
         $this->setTarget(null);
         $this->setUri($uri);
         $this->setVisible(true);
+    }
+
+    /**
+     * Add a navigation node.
+     *
+     * @param AbstractNavigationNode $node The navigation node.
+     * @return AbstractNavigationNode Returns this navigation node.
+     */
+    public function addNode(AbstractNavigationNode $node) {
+        $this->index[$node->getId()] = $this->size();
+        $this->nodes[]               = $node->setParent($this);
+        return $this;
+    }
+
+    /**
+     * Clear the navigation nodes.
+     *
+     * @return AbstractNavigationNode Returns this navigation node.
+     */
+    public function clearNodes() {
+        foreach ($this->getNodes() as $node) {
+            $this->removeNode($node);
+        }
+        return $this;
     }
 
     /**
@@ -100,12 +155,35 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
     }
 
     /**
+     * {@inheritdoc}
+     */
+    public function getAlphabeticalTreeNodeLabel() {
+        return $this->getId();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getAlphabeticalTreeNodeParent() {
+        return $this->getParent();
+    }
+
+    /**
      * Get the enable.
      *
      * @return bool Returns the enable.
      */
     public function getEnable() {
         return $this->enable;
+    }
+
+    /**
+     * Get the first navigation node.
+     *
+     * @return AbstractNavigationNode|null Returns the first navigation node in case of success, null otherwise.
+     */
+    public function getFirstNode() {
+        return $this->getNodeAt(0);
     }
 
     /**
@@ -118,12 +196,93 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
     }
 
     /**
+     * Get the id.
+     *
+     * @return string Returns the id.
+     */
+    public function getId() {
+        return $this->id;
+    }
+
+    /**
+     * Get the last navigation node.
+     *
+     * @return AbstractNavigationNode|null Returns the last navigation node in case of success, null otherwise.
+     */
+    public function getLastNode() {
+        return $this->getNodeAt($this->size() - 1);
+    }
+
+    /**
      * Get the matcher.
      *
      * @return string Returns the matcher.
      */
     public function getMatcher() {
         return $this->matcher;
+    }
+
+    /**
+     * Get a navigation node at.
+     *
+     * @param int $position The position.
+     * @return AbstractNavigationNode|null Returns the navigation node in case of success, null otherwise.
+     */
+    public function getNodeAt($position) {
+
+        if ($position < 0 || $this->size() <= $position) {
+            return null;
+        }
+
+        return $this->getNodes()[$position];
+    }
+
+    /**
+     * Get a navigation node by id.
+     *
+     * @param string $id The id.
+     * @param bool $recursively Recursively ?
+     * @return AbstractNavigationNode Returns the navigation node in case of success, null otherwise.
+     */
+    public function getNodeById($id, $recursively = false) {
+
+        if (true === array_key_exists($id, $this->index)) {
+            return $this->getNodeAt($this->index[$id]);
+        }
+
+        if (false === $recursively) {
+            return null;
+        }
+
+        foreach ($this->getNodes() as $current) {
+
+            $found = $current->getNodeById($id, true);
+            if (null === $found) {
+                continue;
+            }
+
+            return $found;
+        }
+
+        return null;
+    }
+
+    /**
+     * Get the navigation nodes.
+     *
+     * @return AbstractNavigationNode[] Returns the navigation nodes.
+     */
+    public function getNodes() {
+        return $this->nodes;
+    }
+
+    /**
+     * Get the parent.
+     *
+     * @return AbstractNavigationNode Returns the parent.
+     */
+    public function getParent() {
+        return $this->parent;
     }
 
     /**
@@ -154,31 +313,49 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
     }
 
     /**
-     * Determines if the node is displayable.
+     * Determines if this node is displayable.
      *
      * @return bool Returns true in case of success, false otherwise.
      */
     public function isDisplayable() {
-        $displayable = $this->enable && $this->visible;
-        if (true === $displayable) {
+
+        if (true === $this->getEnable() && $this->getVisible()) {
             return true;
         }
+
         foreach ($this->getNodes() as $current) {
-            if (false === ($current instanceof AbstractNavigationNode)) {
+            if (false === $current->isDisplayable()) {
                 continue;
             }
-            if (true === $current->isDisplayable()) {
-                return true;
-            }
+            return true;
         }
-        return $displayable;
+
+        return false;
+    }
+
+    /**
+     * Remove a navigation node.
+     *
+     * @param AbstractNavigationNode $node The navigation node.
+     * @return AbstractNavigationNode Returns this navigation node.
+     */
+    public function removeNode(AbstractNavigationNode $node) {
+
+        if (false === array_key_exists($node->getId(), $this->index)) {
+            return $this;
+        }
+
+        unset($this->nodes[$this->index[$node->getId()]]);
+        unset($this->index[$node->setParent(null)->getId()]);
+
+        return $this;
     }
 
     /**
      * Set the active.
      *
      * @param bool $active Active ?
-     * @return NavigationNode Returns the navigation node.
+     * @return AbstractNavigationNode Returns this navigation node.
      */
     public function setActive($active) {
         $this->active = $active;
@@ -188,8 +365,8 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
     /**
      * Set the enable.
      *
-     * @param bool $enable Enable ?.
-     * @return NavigationNode Returns the navigation node.
+     * @param bool $enable Enable ?
+     * @return AbstractNavigationNode Returns this navigation node.
      */
     public function setEnable($enable) {
         $this->enable = $enable;
@@ -200,7 +377,7 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
      * Set the icon.
      *
      * @param string $icon The icon.
-     * @return NavigationNode Returns the navigation node.
+     * @return AbstractNavigationNode Returns this navigation node.
      */
     public function setIcon($icon) {
         $this->icon = $icon;
@@ -208,10 +385,32 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
     }
 
     /**
+     * Set the id.
+     *
+     * @param string $id The id.
+     * @return AbstractNavigationNode Returns this navigation node.
+     */
+    protected function setId($id) {
+        $this->id = $id;
+        return $this;
+    }
+
+    /**
+     * Set the index.
+     *
+     * @param array $index The index.
+     * @return AbstractNavigationNode Returns this navigation node.
+     */
+    protected function setIndex(array $index) {
+        $this->index = $index;
+        return $this;
+    }
+
+    /**
      * Set the mather.
      *
      * @param string $matcher The matcher.
-     * @return NavigationNode Returns the navigation node.
+     * @return AbstractNavigationNode Returns this navigation node.
      */
     public function setMatcher($matcher) {
         $this->matcher = $matcher;
@@ -219,10 +418,32 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
     }
 
     /**
+     * Set the navigation nodes.
+     *
+     * @param AbstractNavigationNode[] $nodes The navigation nodes.
+     * @return AbstractNavigationNode Returns this navigation node.
+     */
+    protected function setNodes(array $nodes) {
+        $this->nodes = $nodes;
+        return $this;
+    }
+
+    /**
+     * Set the parent.
+     *
+     * @param AbstractNavigationNode $parent The parent.
+     * @return AbstractNavigationNode Returns this navigation node.
+     */
+    protected function setParent(AbstractNavigationNode $parent = null) {
+        $this->parent = $parent;
+        return $this;
+    }
+
+    /**
      * Set the target.
      *
      * @param string $target The target.
-     * @return NavigationNode Returns the navigation node.
+     * @return AbstractNavigationNode Returns this navigation node.
      */
     public function setTarget($target) {
         $this->target = $target;
@@ -230,10 +451,10 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
     }
 
     /**
-     * Set the uri.
+     * Set the URI.
      *
      * @param string $uri The URI.
-     * @return NavigationNode Returns the navigation node.
+     * @return AbstractNavigationNode Returns this navigation node.
      */
     public function setUri($uri) {
         $this->uri = $uri;
@@ -244,10 +465,19 @@ abstract class AbstractNavigationNode extends AbstractNode implements Navigation
      * Set the visible.
      *
      * @param bool $visible Visible ?
-     * @return NavigationNode Returns the navigation node.
+     * @return AbstractNavigationNode Returns this navigation node.
      */
     protected function setVisible($visible) {
         $this->visible = $visible;
         return $this;
+    }
+
+    /**
+     * Size.
+     *
+     * @return int Returns the size.
+     */
+    public function size() {
+        return count($this->getNodes());
     }
 }